cross-rpm: fail early and clearly if needed resources not present
Detect if rpm tar file and patches are missing and give clear message
Fail in rpm build-rpm.sh if anything unexpected happens
Don't create $SDK/rpm dir until it looks pretty certain that build has worked
Change Makefile target so it won't continue on partially built $SDK/rpm dir

+++ cross-rpm/build-rpm.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 
 set -x
+set -e
 
 apply_patch () {
     for i in `echo $1/*.gz `; do
@@ -19,16 +20,22 @@ apply_patch () {
 # Where RPM patches and tarball are stored
 PATCHES=$LINUX_C6X_TOP_DIR/projects/packages/rpm-4.0.4/patches/
 RPM_SRC_DIR=$LINUX_C6X_TOP_DIR/projects/package-downloads/
+RPM_FILE=$RPM_SRC_DIR/rpm-4.0.4.tar.gz
 
-rm -rf $SDK_DIR/rpm
-mkdir -p $SDK_DIR/rpm
+if [ ! -r $RPM_FILE ] ; then
+    echo "$RPM_FILE not found"
+    exit 2
+fi
+if [ ! -d $PATCHES ] ; then
+    echo "Patch directory $PATCHES not found"
+    exit 2
+fi
 
-cp -r cross-rpm $SDK_DIR/rpm/
-
-cd $LINUX_C6X_BUILD_DIR
+mkdir -p $LINUX_C6X_BUILD_DIR
+pushd $LINUX_C6X_BUILD_DIR
 
 rm -rf rpm-4.0.4
-cat $RPM_SRC_DIR/rpm-4.0.4.tar.gz | tar zxf -
+cat $RPM_FILE | tar zxf -
 cd rpm-4.0.4
 
 apply_patch $PATCHES
@@ -36,6 +43,13 @@ apply_patch $PATCHES
 ./configure --prefix=$SDK_DIR/rpm/ --enable-shared=no --enable-static=yes --disable-aio --without-python --without-javaglue
 
 make
+
+rm -rf $SDK_DIR/rpm || true
+mkdir -p $SDK_DIR/rpm
+
 make install
 
+popd
+cp -r cross-rpm $SDK_DIR/rpm/
 
+echo "Build RPM complete" >$LINUX_C6X_BUILD_DIR/rpm-done.txt
